Uploaded car.json
I demonstrate object oriented programming in JSON using my car blueprint that I use in some of my other projects. The is only to store data, such as a C struct, with no member functions. To see an example of using this data in a project, visit my other repo, ObjectOrientedJSONWebDevelopment.